sagemaker_retrieval/requirements.txt
numpy<2
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers>=2.6.0
